Add GitHub action to run linter and build
I want some basic checks to run before allowing merges. `npm test` is
failing with a bunch of things that I don’t have the time to
investigate, presumably resulting from the deletions in 92b2bc4. So for
now just run the linter and make sure that the site builds.