Move date pattern example further down the page
There is more than one way to ask for a date, and that using a date input is only one of the possible approaches, so position the date input example closer to the use case to make this clearer.

This also helps to differentiate the 'asking for dates' pattern from the date input component.