Clarify when you should use inputmode="number"
Change the first sentence from 'If it’s likely that the user will need to enter a number' to 'If you're asking the user to enter a whole number '.

Using inputmode will mean that users relying on virtual keyboards will _only_ be able to enter a number, and so you shouldn't use it in situations where it's only 'likely' the use will need to enter a number.