Add guidance on asking for numbers with decimals
When asking for numbers with decimal places, it's important to use inputtype="decimal" as inputtype="number" may trigger a keyboard that does not include a period key.

The pattern attribute is not required for numbers with decimals – it's used to trigger the numeric keypad on iOS < 12.2, but as far as we can tell there's no equivalent to trigger the keyboard for decimal numbers. On these devices it'll fall back to displaying the full alphanumeric keypad.