Use PaaS-supplied nginx buildpack
When we set the Design System up, PaaS did not support the nginx buildpack, so we had to reference the buildpack using its GitHub URL, pinning to a specific version.

PaaS now provide an nginx buildpack, so we can now use that instead. This means whenever PaaS update their buildpacks we'll automatically use the latest version when we next deploy.

The current version of the buildpack is 1.1.14, so this effectively bumps the buildpack version from 1.1.9 to 1.1.14, and the nginx version from v1.17.10 to v1.19.2 (verified by deploying a test app to the sandbox).