Add scope clause to accessibility statement
Help users understand what this accessibility statement applies to by adding a clause that outlines the scope.

This is based on the sample accessibility statement available on GOV.UK [1]. The scope clause is listed as part of the ‘mandatory content’ in the model statement published by the EU [2], on which the sample accessibility statement is based.

[1]: https://www.gov.uk/government/publications/sample-accessibility-statement/sample-accessibility-statement-for-a-fictional-public-sector-website

[2]: https://ec.europa.eu/info/law/better-regulation/have-your-say/initiatives/1753-Commission-implementing-decision-on-establishing-a-model-accessibility-statement-under-the-WAD#:~:text=Annex%20-%20Ares(2018)2604172