Associate WCAG acronym with expanded form
We use both Web Content Accessibility Guidelines and the WCAG acronym in this document, but we don’t currently associate the two.

Include the WCAG acronym after the first use of its expanded form, so that users understand that they are related to each other.

Use the acronym consistently once the term has been introduced.