Update ID and name for conditional reveal examples
The conditional reveals example for the checkboxes component uses an ID of `citizenship-conditional` but the question is about contact preferences – suspect this is either a copy and paste error or an error introduced when the content in the example was updated. Update it to use ‘contact’ as a prefix instead.

Update the corresponding example in the radios component to be consistent rather than the current more verbose `how-contacted-conditional`.