Update redirect pages
At the minute we try to construct a sentence based on the page title, which is both the title for the archived page and the name of the page we link the user to (for example, “You can view the up to date information about [How you can contribute].”)

Change the way we archive pages to allow the body to be included in the page, which allows us to write a message that makes sense for the page being archived.

This also gives us the option of using the archived layout without ‘redirecting’ the user to another page, in case we need to archive a page that does not have a logical place to redirect to.

Update the page title to use the original page title suffixed by ‘(archived)’ rather than ‘page has been archived’, for example ‘How you can contribute (archived)’ rather than ‘How you can contribute page has been archived’.