Exclude archived pages from navigation
We can use `ignore_in_sitemap` as a proxy for ‘archived’ as if something is excluded from the sitemap it makes sense that it shouldn’t appear in the navigation either.

We’ll need to be able to support this as we look to retire the ‘Help users to understand the impact of an emergency on your service’ pattern after the navigation banner component is released.