Treat hexadecimal colours codes as code
They are snippets of code, so semantically it makes sense to treat them
as code.

This also has the benefit of styling them as code, which makes the
individual digits easier to distinguish.

I’ve overridden the red text colour and grey background so they are
styled the same as the colours’ variable names.