Improve visibility of focus state for search box
Prompted by discussions with the NHS Frontend team (https://github.com/nhsuk/nhsuk-frontend/issues/679), improve the visibility of the focus state for the search box by adding a black (#0b0c0c) border between the white text input and the yellow focus outline.

Without this change, the only part of the focus indicator that meets colour contrast is the outermost 1px which change from black to white as the input ‘grows’ – the other 2px of the 3px outline change from white to black.

After this change, the entire 7px comprised of the border, box shadow and outline meet contrast, as the border changes from a 2px white to a 4px (2px border + 2px box-shadow) black (19.58), and the 3px outline causes a change from black to yellow (14.54).