Render additional component descriptions as markdown
We render macro options as markdown so things like backticks get rendered as <code></code>.
However, we weren't converting the descriptions to markdown for additional components, e.g: the label param within the text input component. This meant some text displayed as "If `html` is set, this is not required" instead of "If <code>html</code> is not set, this is not required".