Exclude ethnic group pattern from sitemap and nav
The ‘asking users for ethnic groups’ pattern was archived in #1582 [1] and redirected to the new ‘asking users for equality information’ pattern.

However, the Design System currently relies on the ignore_in_sitemap flag in the frontmatter when deciding whether the page should appear in the sitemap and the site navigation.

This means that the pattern is still appearing in the sitemap, and also at the very bottom of the Patterns navigation.

Add the `ignore_in_sitemap` flag to the archived page to remove it from both places.

[1]: https://github.com/alphagov/govuk-design-system/pull/1582