Enable the new link and hover styles
The new link and hover styles in GOV.UK Frontend v3.12.0 will be opt-in until the next major
release of GOV.UK Frontend. However, we'd like the Design System website itself to take
advantage of the new styles when 3.12.0 is released in order to present the most up-to-date
version of GOV.UK Frontend.