Make Equality Info example question accessible
What
----
Addresses [#1624](https://github.com/alphagov/govuk-design-system/issues/1624).

This PR changes an example question in the [Equality information pattern](https://design-system.service.gov.uk/patterns/equality-information/) from:

> Which of the following options best describes how you think of yourself?

to

> Which of the following best describes your sexual orientation?

Why
---
This change is because of an issue raised by a designer at the Office of National Statistics (ONS). The original question was inaccessible - when screen readers read it out, what was being asked was unclear to users.

The revised wording is in line with what was asked in the Census and signed off by the ONS.