Add copy of cookie-functions JavaScript module from Digital Marketplace
This commit adds copies of the [cookie consent functions from
digitalmarketplace-govuk-frontend][1].

We wanted a cookie module to provide the following functionality:

- accept and reject cookies
- save consent in a cookie, with an expiry of 1 year
- delete cookies when a user changes their preference to reject
- look up a user's consent
- be able to map a cookie name to it's category, e.g: `_ga` -> `analytics`

We reviewed existing code in other alphagov projects and found the
Digital Marketplace GOV.UK Frontend code fit all of our needs and was
the best fit for our existing codebase. See issue #1615 for more details
of the analysis done.

This commit copies `cookie-functions.js` and `cookie-functions.test.js`
to `src/javascripts` where it can be used by future code for the cookie
banner and cookie page.

[1]: https://github.com/alphagov/digitalmarketplace-govuk-frontend/tree/d24e7f1dc6e94fb9f173aebf46aed5f1051af59c/src/digitalmarketplace/helpers/cookie