Add ECMAScript Modules transpiler to dev dependencies for tests
We want to be able to unit test some of our JavaScript modules, however
we write our scripts using the ES6+ import and export statements, which
are not yet fully supported by the Jest test runner we use. One
recommended workaround is to use the Babel transpiler to process the
script[[1]].

This commit adds the Babel package needed to compile ES6 modules to
CommonJS modules, and configures it using a `.babelrc.json` file so that
it is used only when Jest is running.

Note that this is only needed for tests, for code used in production and
for browser tests we use the plugin `rollup-plugin-commonjs` to do a
similar thing. Unfortunately we can't use rollup for tests, because the
library [rollup-jest][2] that would support this requires rollup ^1.1.0,
and `metalsmith-rollup` requires rollup ^0.5.8.

[1]: https://stackoverflow.com/questions/35756479/does-jest-support-es6-import-export
[2]: https://github.com/ambar/rollup-jest