Refactor deleting of cookies
The refactor in the previous commit exposed a bug where the `setCookie`
function was setting a cookie to `null` instead of deleting it.
Unfortunately the test was not picking this up. This commit fixes the
test, and adds a function explicitly for deleting cookies, which is
simpler than trying to get the right arguments for `setCookie`.

This makes the implementation of `Cookie` more explicit, and also allows
us to simplify the implementation of `checkConsentCookie`, because we
no longer need to check the user consent when deleting a cookie.