Refactor cookie functions to make essential cookies explicit
This commit refactors the cookie functions with the aim of making it
clearer how essential cookies should be handled.

Essential cookies are essential to the running of the site (either the
site breaks or a large portion of it doesn't work without cookies).

What this means for the consent functions is that we don't want to allow
users to deslect essential cookies using the cookie consent mecahnism
(note that they can still choose to reject all cookies via browser
configuration).

This means that we shouldn't include an option for essential cookies in
the consent preferences cookie, and shouldn't act if such an option is
somehow present.

This commit makes this clear with both implementation and comments.