Change behaviour of cookie functions when user does not have preferences
We want our cookies to be opt-in, so if there is no consent cookie set
we shouldn't set any cookies (except essential ones).

This commit changes the cookie functions so that when checking if a
cookie can be set, if the user has not set any preferences it checks
against DEFAULT_COOKIE_CONSENT.

This means that, since our default preferences cookie disallows all
cookie categories, cookies are now opt-in.

This commit also updates the tests to reflect this.