Add what's new section and signup to homepage
We want a way for users of the Design System website to find out about
changes to guidance or about the addition of new components or patterns.

This commit adds a 'what's new' section to the homepage, with details
about the latest release, and a link to sign up to the 'what's new'
email updates.

This design is based on user research performed with commit 76fe0ad.
Currently the email signup journey is implemented using a Mailchimp
landing page; in future a hosted form may be used (see issue #1731).

Co-authored-by: Calvin Lau <77630796+calvin-lau-sig7@users.noreply.github.com>