Add versioning to consent cookie
We want to be able to invalidate users' cookie consent preferences in
the scenario where we've changed the policy, or the cookie format.

This commit adds a version number to the cookie as an additional
property.

In the case where a user has a version number which differs to the
expected version, then `getConsentCookie` will still return the
cookie preferences (with the old version number), so that you can show
the old cookie preferences to the user when asking for new preferences.

However, you cannot set cookies with `Cookie` until the cookie
preferences have been updated, and as a consumer of the library should
avoid doing anything involving cookies if the version is incorrect. This
commit also adds a function `isValidConsentCookie` to test whether the
consent cookie object has the correct version.