Add resetCookies function
In the scenario where a user has an out of date consent cookie, we may
want to delete all the cookies they have previously consented to.

This commit adds a function `resetCookies`, which synchronises the
user's cookie preferences with their cookies. If there are any cookies
present belonging to a category which the user has not consented to,
then those cookies are deleted. If the user has consented to cookies
previously but has an outdated consent cookie, then `resetCookies` will
use the default cookie preferences instead of the users cookie
preferences, which results all opt-in cookies being deleted.