Add border to bottom of masthead and remove border from top of 'what's new'
We want a divider to appear between the masthead and the what's new
section when the user is using high contrast mode or similar.

This commit adds a 1px blue border to the bottom of the masthead;
without high contrast mode the border is indistinguishable from the
masthead background, but if high contrast mode is enabled (or some other
forced colour mode that removes backgrounds), the border will be
visible.

Co-authored-by: Charlotte Downs <56260216+CharlotteDowns@users.noreply.github.com>