Add cookie preferences form
This commit adds a "form" to the cookies page for users to be able to
update their cookie consent preferences.

As the Design System website is a static site without a backend to
process form submissions, this page uses JavaScript to process the form
input, show notification banners, and update the cookies.  We
only use cookies with JavaScript, so it is acceptable for users to not
be able to update their cookie preferences if they do not have
JavaScript enabled. In the case where JavaScript is not available the
form is not visible.

The cookie page has several dynamic components that are controlled by
JavaScript:

- a success notification banner which is shown only after cookie
settings have been saved

- a set of radios for each cookie type

- a button which is used to submit the form

Currently only one cookie type is supported, analytics, but the
JavaScript in `src/javascripts/components/cookies-page.js` should be
flexible enough to be expanded to support multiple cookie types without
too much additional code.