Mention alternatives to "none of these"
Add a sentence to Checkboxes components to tell service teams to consider alternatives to the 'None of these' component, as it may not be the best option for some questions.

Community and working group says that it is important to acknowledge that there are several ways of dealing with the 'none of the above' case and we ought to urge designers to consider different options.

Content designer, Designer

Discussions in the backlog for [Checkboxes](https://github.com/alphagov/govuk-design-system-backlog/issues/37) component.

- [ ] Draft sentence
- [ ] 2i with team, check with people concerned
- [ ] Publish

Change guard question to filter question

It might be clearer to call these filter questions