Align the phase banner content to app-width-container on mobile
The content within the phase banner on mobile has left and right padding applied.
This means the content within the banner does not line up with other content on the
page that is within an app-width-container element.

The app-width-container class means the padding isn't actually needed, so we can
remove it on mobile. There is already no padding on desktop.