Replace const with var
We want to support IE10 and below, which do not support the ES6 `const`
keyword.

I thought our uglifier (metalsmith-uglify) would remove const,
but it does not. We would need an additional tool to do so.

So rather than making our system more complicated, lets just get rid of
the const declarations.