Change width override example to be more generic
The width override example at the moment uses a 'full name' text input. This doesn't match the example
used in the text input guidance and also means we should probably include the autocomplete attribute which
may lead to more copy/paste errors from users.

Instead, change the example to be more generic and to match the existing text input guidance.