Remove border from banner on homepage
The bottom border on the banner ends up butting up against the top of the masthead on the homepage at the mobile viewport. As it’s inset with the grid gutter it does not span the full width of the masthead, so this looks a bit odd.

Remove the bottom border from the phase banner from the homepage page only, by setting a `masthead` boolean in the frontmatter.