Add sentence about IE8 and previous versions
The cookie banner and cookies form is hidden on IE8 and below.
We want to add a sentence to the cookies page when users are using these versions,
to explain that the information on the page doesn't apply to them.

Note: we're hiding the cookie banner and cookies form on IE8 and below,
because the Google Tag Manager script is officially not supported on these versions.