Set jest testURL to https://design-system.service.gov.uk
Without this config, the default URL is 'localhost' which causes
some issues when setting cookies because of a bug in tough-cookie which treats
localhost like a public domain. Because you can't set cookies on a public domain,
cookies were not being set in the tests.

Setting the testURL to a real URL fixes this issue.

@lfdebrux has raised an issue with tough-cookie here: https://github.com/salesforce/tough-cookie/issues/215