Disable/enable google analytics depending on consent
Deleting the cookies that Google Analytics sets is not enough to disable analytics,
especially if you continue to navigate on the same page in which you turned them off -
for our users this is probably the cookies page.

Instead, we need to tell Google Analytics to stop tracking/firing events using the
window['ga-disable-ua-XXXXX-Y'] window property which it checks each time ga() is called.
This will prevent any further events being sent to Google Analytics.

As part of this, we also need to enable Google Analytics if the reverse happens (i.e:
a user consents to analytics)

Google documentation: https://developers.google.com/analytics/devguides/collection/analyticsjs/user-opt-out