Generalise method for deleting test cookies
Now that we have more cookies being allowed/set in cookie-functions
due to the addition of _gat cookies, this commit adjusts the logic we have
in the afterEach() section of the cookie-functions tests so that we no longer
have to specify the name for each cookie we want to delete (which we do to avoid
tests interfering with other tests).

Instead, we can simply loop through all cookies that have been set and delete them.
We shouldn't need to rely on any cookies existing from one test to the next - delete them all.