Replace use of FormElement.elements to support IE
The MDN page for FormElement.elements states that it is fully supported by IE.
However, in IE, FormElement.elements returns a HTMLCollection rather than a RadioNodeList, which means
we can't use RadioNodeList.value to check or set the radio button values.

This was leading to a bug where the radio buttons in the cookies forms were not being
preselected, and the values were not being correctly set.

Instead of relying on the elements being returned as a RadioNodeList:

- use `querySelector` and the `:checked` selector to find radio buttons which have been selected already
- use `querySelector` and the `[name]` and `[value]` selectors to find the right radio button to select