Delete cookies with and without domain
Cookies need to be deleted in the same level of specificity in which they were set.
You can't tell if a cookie was set with a domain attribute or not, so try both options

Previously, we set the Google Analytics cookieDomain property to 'auto' which caused
an issue in Internet Explorer where the cookies were set on service.gov.uk. We then set
the cookieDomain property to 'design-system.service.gov.uk' or the output of {{Page Hostname}}
in Google Tag Manager for preview apps. However, this actually means the analytics cookies are
set on the specified domain *and any of it's subdomains*, which isn't the behaviour we want at
the moment.

So we now set the cookieDomain property to "none", which effectively sets the analytics cookies
without any specified domain, meaning the cookies apply to that domain and that domain only. This
is the behaviour we want, but it means our deletion code doesn't work because we're specifying the domain.
This commit changes the deleteCookie function to handle both - this way, if we make changes to the
cookieDomain property again in the future via Google Tag Manager, the delete code should continue
to work.