Update deploy job to use production environment
Part of https://github.com/alphagov/design-system-team-internal/issues/446.

We want to make sure only merges to main can deploy to PaaS, so we've
created a [GitHub Actions environment][1] that contains the production
secrets with the necessary protection rules.

This commit sets the job environment to the new `production`
environment, so the job can use the new secret.

[1]: https://docs.github.com/en/actions/reference/environments