Add govuk-body class to design system cookie banner
Previously we were relying on default colour and margins for the text content within the cookie banner,
when html was passed to the banner. In the yaml examples in govuk-frontend, we assume a user will
add govuk-body classes to any html they pass in. However, this was missed when adding the cookie banner
to the design system site. This commit adds the missing govuk-body classes.