Add archive banner
We want it to be clear to users that when they visit
v2--govuk-design-system-preview.netlify.com that what they are seeing is
an old version of the Design System website.

This commit changes the phase banner for [Netlify branch deploys][1] to
state that the website is an archive (rather than a preview).

Note that this applies to all branch deploys, so it might potentially
cover branches that are for pull requests, but we only usually share
branch deploys for old versions of the website.

[1]: https://docs.netlify.com/site-deploys/overview/