Disallow redirects when checking for broken links
Many of the links to other pages in the Design System currently omit the trailing slash.

When you omit the trailing slash from a URL in the Design System, the web server will redirect you to the version with the trailing slash – for example if you try to go to design-system.service.gov.uk/components you'll be redirected to design-system.service.gov.uk/components/.

This adds an extra round trip to the server which could otherwise be avoided if we ensure that all links include the trailing slash.

Avoiding the unnecessary redirect will improve performance, especially for users on slower connections.

For example, on a simulated slow 3G connection, the extra round trip for the 301 redirect takes approx 2 seconds, out of a total 5.77s (till DOMContentLoaded).

The performance impact is still relatively worthwhile but is objectively less noticeable for users on a fast connection – on a 120mb fibre connection the request takes 60ms out of a total 300ms (till DOMContentLoaded).

Configure the broken link checker to disallow redirects, which will cause it to fail if any of the links it find result in a redirect.