Cache npm dependencies between builds
Use the built-in functionality provided by the setup-node GitHub Action[1] to cache and restore dependencies.

This will enable quicker builds[2] by caching the node_modules folder between builds, as long as the package-lock.json file hasn't changed (a hash of the file is used as part of the cache key).

We also need to bump to v2.x of setup-node, as caching dependencies was introduced in v2.2.0[3].

[1]: https://github.com/actions/setup-node#caching-packages-dependencies
[2]: https://docs.github.com/en/actions/advanced-guides/caching-dependencies-to-speed-up-workflows
[3]: https://github.com/actions/setup-node/tree/v2.2.0