Add a full page GOV.UK example layout
Our current full page example layout contains a header with a service name.
We're using this layout for examples of pages on GOV.UK, even though the GOV.UK header
doesn't use a service name.

This commit creates a new full page layout specific to GOV.UK. At the moment, the only
difference is that it doesn't use a service name in the header. However, there might be
other changes we have to make in future if GOV.UK diverge from the design system and/or
make use of different component options than services.