Add CI workflow step to prevent deployment of old code
We want to prevent against the GitHub Actions CI workflow being run in a
circumstance that would result in old/stale/out-of-date code being
deployed to production.

This commit adds a new step to the workflow that must pass before a
deploy can take place. It contains a single check currently, which uses
JavaScript to check the commit sha of the code that would be deployed
against the most recent commit sha of the main branch. If this check
fails, the user who started the workflow run will be notified of the
failure, but we add a warning and a message explaining what has happened
and asking them not to try re-running (otherwise they might end up
trying to re-run it lots of times).