Add upload-artifact input
Changes the test workflow so that rather than only publishing an
artifact when run on the main branch, it is an input argument of the
workflow call [[1]].

The commit adds an input that specifies whether to upload a zip of the
build folder as an artifact. We usually only want to do that when we're
doing a deploy, because we don't want to run setup and build twice, and
we want the deploy to be a separate job, and you can't share folders
between jobs any other way. I thought it was a bit of weird magic that
the test job would upload if the branch was main, so I thought making
this change would make the link between that behaviour and deploy jobs
more explicit.

This makes testing of the deploy workflow slightly easier, and also
reduces the number of places we have to encode the name of the main
branch.

[1]: https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#onworkflow_callinputs