Update workflow_call ref for test.yaml
Updates the ref for the workflow call in `deploy.yaml`.

GitHub Actions currently requires a ref for workflow calls [[1]], which
has to be a commit SHA, a release tag, or branch name. We can't use the
branch name `main` because the addition of `test.yaml` hasn't been
merged yet, and we don't want to have to manage a branch name or release
tag, so we use a commit SHA.

This has to be a separate commit because otherwise we wouldn't know the
commit SHA that adds `test.yaml`.

[1]: https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#jobsjob_iduses