Add concurrency limit to production deploys
Refactors deploy workflow to use a concurrency limit to avoid trying to
run more than one deploy at a time [[1]].

This is an alternative to the step which cancels any in-progress
deploys, which would have resulted in failed deployment messages.

This should make it easier to spot when a deployment fails for real.

[1]: https://github.blog/changelog/2021-04-19-github-actions-limit-workflow-run-or-job-concurrency/