Mention PaaS in our Logit documentation
We mention the Reliability Engineering team in our Logit docs, but they're not
able to help with questions about how our data gets sent from the application through
to Logit.

For this, we should be able to ask the PaaS team who can answer our questions or
point us in the right direction.

This came up as part of investigating whether we needed to take action
for a Logit certification rotation https://github.com/alphagov/design-system-team-internal/issues/544