Move cookie banner before skip link
In the design system guidance for the cookie banner, we tell users to
place the cookie banner before the skip to main content link using the
`bodyStart` block.

However, we didn't follow our own guidance when adding the cookie banner to
the design system website - the cookie banner is placed after the skip link.

We had a report on Zendesk that this makes it difficult for a user to access
the skip link once they've accepted/rejected cookies and hidden the cookie banner,
because the focus carries on from the cookie banner rather than returning to the
top of the page. To get back to the skip link, they need to know to navigate backwards
or cycle all the way back through the page.

We should follow our own guidance, and place the cookie banner before the skip link.

Our research into positioning of the cookie banner: https://github.com/alphagov/govuk-frontend/issues/2096#issuecomment-767667620