Avoid repetition in banner logic
Set the tag and banner text depending on the PULL_REQUEST, then use them in a single call to govukPhaseBanner.

Remove the word 'phase' from all variables as we're using the phase banner in an unconventional way which is now never going to relate to the phase the Design System service is in.