Use 'Address' in summary list examples
At the minute we're using two very similar terms ('contact information' and 'contact details') to refer to two different things. This was previously flagged by DAC as confusing and failing WCAG 2.1 2.4.9: Link Purpose (Link Only) (because of the visually hidden text in the links – see #906).

This mirrors the change previously made to similar examples in the 'Help users check their answers' pattern as part of 36b3475 (#912).