Ensure character count id is unique on the page
We've made changes in GOV.UK Frontend v4.0.0 so that the JS now
looks through the whole page rather than just within the module.
This means we need to make sure the `id` for character counts on
the same page are unique.

I don't think this should be an issue as we put our examples within iframes,
but it's good to be consistent