Use govuk template layout for examples
The example layouts in `layout-example.njk` were using a hardcoded layout
instead of using the template.

Switch the layout to use the template.  This means we need to block out the
header, skip link and footer as we don't want anything else on the example pages.

This means we get things like meta tags and social sharing images "for free" without
having to remember to update anything hardcoded.