Make navigation links consistent with other links
Our ‘standard’ links turn darker blue on hover. This CSS is currently overriding that behaviour so that the links _do not_ change colour on hover.

This was discussed when we first introduced the new link styles to the Design System [1]. It seems like we were mainly looking to preserve the existing behaviour, but it’s not clear that this is desirable and nobody knew of any reason why the links shouldn’t get darker on hover.

Remove the overrides so that the navigation links are consistent with other links across the Design System.

[1]: https://github.com/alphagov/govuk-design-system/pull/1648#discussion_r630122484