Remove ajv from direct dependencies
The commit 1aedd1a59 says that we added it because we needed it as a
peer-dependency of `ajv-keywords`, but we don't have `ajv-keywords` in
our package file, it's a fourth-level sub-dependency.

```
% npm ls ajv-keywords
govuk_design_system
├─┬ sass-lint@1.13.1
│ └─┬ eslint@2.13.1
│   └─┬ table@3.8.3
│     └── ajv-keywords@1.5.1
└─┬ standard@12.0.1
  └─┬ eslint@5.4.0
    └─┬ table@4.0.3
      └── ajv-keywords@3.2.0
```

I'm pretty sure the way peer dependencies work is that you're supposed
to require both at the same level, and indeed `table` (which requires
`ajv-keywords` in our dependency tree). So we should be able to remove
`ajv` from `package.json` without this error.

I seem to recollect that in the past `npm` had a bug where it would warn
about peer dependencies erroneously, but I can't find a reference to it
right now.