Replace separate navs for desktop and mobile with one
Having a single nav element that changes is better for progressive enhancement (no CSS specifically) and web performance.

This change additionally replaces class and aria-hidden as a means to hide elements with the native hidden attribute which is more reliable as a means to hide with CSS off and for assistive tech.

Majority of this implementation borrowed from current solution for gov.uk header navigation.