Add missing class to cookie banner in page template
Following on from f85db87 (#1891), we also need to add the govuk-body class to the paragraphs in the cookie banner in the ‘custom page template’ example.

Previously we were relying on default colour and margins for the text content within the cookie banner, when html was passed to the banner. In the yaml examples in govuk-frontend, we assume a user will add govuk-body classes to any html they pass in.

We then went on to remove the default font styles from the cookie banner in 4.0.0 [1], to make it more obvious when you have not added classes and styles to any custom HTML – but didn’t notice we had an example in the Design System which did just that!

[1]: https://github.com/alphagov/govuk-frontend/blob/main/CHANGELOG.md#style-any-custom-html-in-your-cookie-banner