Link to scala implementation of govuk-frontend
Was mentioned in cross gov slack that there wasn't a link to this library in the resources and tools section so I volunteered to propose a change to list it. This library is called play-frontend-hmrc but it also contains an implementation of all govuk components. This library is maintained by the Platform UI team on the HMRC Tax Platform.