Use text option for address pattern example
We previously included some HTML in the label for these two fields in order to add visually hidden suffixes.

We try to ensure we use the `text` option over the `html` option unless passing HTML, as the `text` option escapes any user input that might be passed as part of the string, preventing XSS attacks.

On that basis, now that these labels are ‘just text’, update the example to use the `text` option.